+++ src/App.tsx
@@ -1,27 +1,27 @@
 import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator } from "@react-navigation/stack";
 import React from "react";
-import { AccelerometerScreen } from "./AccelerometerScreen";
-import { AmplitudeScreen } from "./AmplitudeScreen";
-import { AssetScreen } from "./AssetScreen";
-import { AudioScreen } from "./AudioScreen";
-import { BarCodeScannerScreen } from "./BarCodeScannerScreen";
-import { BlurViewScreen } from "./BlurViewScreen";
-import { BrightnessScreen } from "./BrightnessScreen";
-import { CameraScreen } from "./CameraScreen";
-import { ConstantsScreen } from "./constants/ConstantsScreen";
-import { ManifestScreen } from "./constants/ManifestScreen";
-import { PlatformScreen } from "./constants/PlatformScreen";
-import { SystemFontsScreen } from "./constants/SystemFontsScreen";
-import { FacebookScreen } from "./FacebookScreen";
-import { FontScreen } from "./FontScreen";
-import { GyroscopeScreen } from "./GyroscopeScreen";
+import { AccelerometerScreen } from "./demoScreens/AccelerometerScreen";
+import { AmplitudeScreen } from "./demoScreens/AmplitudeScreen";
+import { AssetScreen } from "./demoScreens/AssetScreen";
+import { AudioScreen } from "./demoScreens/AudioScreen";
+import { BarCodeScannerScreen } from "./demoScreens/BarCodeScannerScreen";
+import { BlurViewScreen } from "./demoScreens/BlurViewScreen";
+import { BrightnessScreen } from "./demoScreens/BrightnessScreen";
+import { CameraScreen } from "./demoScreens/CameraScreen";
+import { ConstantsScreen } from "./demoScreens/constants/ConstantsScreen";
+import { ManifestScreen } from "./demoScreens/constants/ManifestScreen";
+import { PlatformScreen } from "./demoScreens/constants/PlatformScreen";
+import { SystemFontsScreen } from "./demoScreens/constants/SystemFontsScreen";
+import { FacebookScreen } from "./demoScreens/FacebookScreen";
+import { FontScreen } from "./demoScreens/FontScreen";
+import { GyroscopeScreen } from "./demoScreens/GyroscopeScreen";
+import { LinearGradientScreen } from "./demoScreens/LinearGradientScreen";
+import { LocalAuthenticationScreen } from "./demoScreens/LocalAuthenticationScreen";
+import { MapViewScreen } from "./demoScreens/MapViewScreen";
+import { SvgScreen } from "./demoScreens/SvgScreen";
+import { VectorIconsScreen } from "./demoScreens/VectorIconsScreen";
 import { HomeScreen } from "./HomeScreen";
-import { LinearGradientScreen } from "./LinearGradientScreen";
-import { LocalAuthenticationScreen } from "./LocalAuthenticationScreen";
-import { MapViewScreen } from "./MapViewScreen";
-import { SvgScreen } from "./SvgScreen";
-import { VectorIconsScreen } from "./VectorIconsScreen";
 
 const Stack = createStackNavigator();
