+++ src/demoScreens/AccelerometerScreen.tsx
