+++ src/demoScreens/AmplitudeScreen.tsx
