+++ src/demoScreens/BarCodeScannerScreen.tsx
