+++ src/demoScreens/LocalAuthenticationScreen.tsx
